# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.9.0-rc.1",
    "title": "WorkerCore API"
  },
  "components": {
    "schemas": {
      "ResolveResponse": {
        "type": "object",
        "properties": {
          "i": {
            "type": "string",
            "example": "worker-123"
          }
        },
        "required": [
          "i"
        ]
      },
      "ResolveQuery": {
        "type": "object",
        "properties": {
          "query": {
            "nullable": true,
            "example": {
              "getForId": {
                "workerId": "worker-123"
              }
            }
          }
        }
      },
      "ActionResponse": {
        "nullable": true
      },
      "ActionRequest": {
        "type": "object",
        "properties": {
          "query": {
            "nullable": true,
            "example": {
              "getForId": {
                "workerId": "worker-123"
              }
            }
          },
          "body": {
            "nullable": true,
            "example": {
              "param1": "value1",
              "param2": 123
            }
          }
        }
      },
      "ConnectionMessageResponse": {
        "nullable": true
      },
      "ConnectionMessageRequest": {
        "type": "object",
        "properties": {
          "message": {
            "nullable": true,
            "example": {
              "type": "message",
              "content": "Hello, worker!"
            }
          }
        }
      }
    },
    "parameters": {}
  },
  "paths": {
    "/workers/resolve": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Worker query information"
            },
            "required": true,
            "name": "X-AC-Query",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "400": {
            "description": "User error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/workers/connect/websocket": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The encoding format to use for the response (json, cbor)",
              "example": "json"
            },
            "required": true,
            "name": "encoding",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Worker query information"
            },
            "required": true,
            "name": "query",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket upgrade"
          }
        }
      }
    },
    "/workers/connect/sse": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The encoding format to use for the response (json, cbor)",
              "example": "json"
            },
            "required": true,
            "name": "X-AC-Encoding",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Worker query information"
            },
            "required": true,
            "name": "X-AC-Query",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Connection parameters"
            },
            "required": false,
            "name": "X-AC-Conn-Params",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/workers/actions/{action}": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "myAction"
            },
            "required": true,
            "name": "action",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "The encoding format to use for the response (json, cbor)",
              "example": "json"
            },
            "required": true,
            "name": "X-AC-Encoding",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Connection parameters"
            },
            "required": false,
            "name": "X-AC-Conn-Params",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "User error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/workers/message": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Worker ID (used in some endpoints)",
              "example": "worker-123456"
            },
            "required": true,
            "name": "X-AC-Worker",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Connection ID",
              "example": "conn-123456"
            },
            "required": true,
            "name": "X-AC-Conn",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "The encoding format to use for the response (json, cbor)",
              "example": "json"
            },
            "required": true,
            "name": "X-AC-Encoding",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Connection token"
            },
            "required": true,
            "name": "X-AC-Conn-Token",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "User error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    }
  }
}